+++ test/operator/one_point_crossover_strategy.cpp
@@ -31,7 +31,7 @@ TYPED_TEST( one_point_crossover_strategy_tests , instanciation )
     ASSERT_NO_THROW( 
     for( size_t i=0 ; i<10 ; ++i )
     {
-        auto c = gpcxx::make_one_point_crossover_strategy( this->m_gen.rng , 10 );
+        auto c = gpcxx::make_one_point_crossover_strategy( this->m_gen.rng , size_t( 10 ) );
         size_t l1 = this->m_test_trees.data.size() + this->m_test_trees.data2.size();
         cerr << this->m_test_trees.data.size() << " " << simple( this->m_test_trees.data ) << "\n";
         cerr << this->m_test_trees.data2.size() << " " << simple( this->m_test_trees.data2 ) << "\n\n";
@@ -50,15 +50,15 @@ TYPED_TEST( one_point_crossover_strategy_tests , heights )
     ASSERT_NO_THROW(
     for( size_t i=0 ; i<10 ; ++i )
     {
-        auto c = gpcxx::make_one_point_crossover_strategy( this->m_gen.rng , 5 );
+        auto c = gpcxx::make_one_point_crossover_strategy( this->m_gen.rng , size_t( 5 ) );
         cerr << this->m_test_trees.data.size() << " " << simple( this->m_test_trees.data ) << "\n";
         cerr << this->m_test_trees.data2.size() << " " << simple( this->m_test_trees.data2 ) << "\n\n";
         c( this->m_test_trees.data , this->m_test_trees.data2 );
         cerr << this->m_test_trees.data.size() << " " << simple( this->m_test_trees.data ) << "\n";
         cerr << this->m_test_trees.data2.size() << " " << simple( this->m_test_trees.data2 ) << "\n";
         cerr << endl;
-        EXPECT_LE( this->m_test_trees.data.root().height() , 5 );
-        EXPECT_LE( this->m_test_trees.data2.root().height() , 5 );
+        EXPECT_LE( this->m_test_trees.data.root().height() , size_t( 5 ) );
+        EXPECT_LE( this->m_test_trees.data2.root().height() , size_t( 5 ) );
     }
     );
 }
