+++ test/tree/basic_tree.cpp
@@ -25,18 +25,18 @@ using namespace gpcxx;
 
 #define TESTNAME basic_tree_tests
 
-#define TEST_NODE( n , VALUE , ARITY , HEIGHT , LEVEL )          \
-    EXPECT_EQ( *(n) , VALUE );                                   \
-    EXPECT_EQ( (n).size() , ARITY );                             \
-    EXPECT_EQ( (n).height() , HEIGHT );                          \
-    EXPECT_EQ( (n).level() , LEVEL )
+#define TEST_NODE( n , VALUE , ARITY , HEIGHT , LEVEL )                   \
+    EXPECT_EQ( *(n) , VALUE );                                            \
+    EXPECT_EQ( (n).size() , size_t( ARITY ) );                            \
+    EXPECT_EQ( (n).height() , size_t( HEIGHT ) );                         \
+    EXPECT_EQ( (n).level() , size_t( LEVEL ) )
 
 
 TEST( TESTNAME , default_construct )
 {
     basic_tree< int > tree;
     /* auto root = */ tree.root();
-    EXPECT_EQ( tree.size() , 0 );
+    EXPECT_EQ( tree.size() , size_t( 0 ) );
     EXPECT_TRUE( tree.empty() );
 }
 
@@ -46,7 +46,7 @@ TEST( TESTNAME , insert_value )
     auto root = tree.root();
     basic_tree< std::string >::cursor n1 = tree.insert_below( root , "+" );
     
-    EXPECT_EQ( tree.size() , 1 );
+    EXPECT_EQ( tree.size() , size_t( 1 ) );
     EXPECT_FALSE( tree.empty() );
     TEST_NODE( n1 , "+" , 0 , 1 , 0 );
     TEST_NODE( tree.root() , "+" , 0 , 1 , 0 );
@@ -55,7 +55,7 @@ TEST( TESTNAME , insert_value )
     tree.insert_below( n1 , "12" );
 
     EXPECT_FALSE( tree.empty() );
-    EXPECT_EQ( tree.size() , 3 );
+    EXPECT_EQ( tree.size() , size_t( 3 ) );
     TEST_NODE( tree.root() , "+" , 2 , 2 , 0 );
     TEST_NODE( tree.root().children(0) , "11" , 0 , 1 , 1 );
     TEST_NODE( tree.root().children(1) , "12" , 0 , 1 , 1 );
@@ -72,7 +72,7 @@ TEST( TESTNAME , insert_and_erase )
     /* auto n5 = */ tree.insert_below( n2 , "15" );
     /* auto n6 = */ tree.insert_below( n2 , "16" );
     
-    EXPECT_EQ( tree.size() , 7 );
+    EXPECT_EQ( tree.size() , size_t( 7 ) );
     EXPECT_FALSE( tree.empty() );
     TEST_NODE( tree.root() , "+" , 2 , 3 , 0 );
     TEST_NODE( tree.root().children(0) , "-" , 2 , 2 , 1 );
@@ -84,7 +84,7 @@ TEST( TESTNAME , insert_and_erase )
     
     tree.erase( n2 );
     
-    EXPECT_EQ( tree.size() , 4 );
+    EXPECT_EQ( tree.size() , size_t( 4 ) );
     EXPECT_FALSE( tree.empty() );
     TEST_NODE( tree.root() , "+" , 1 , 3 , 0 );
     TEST_NODE( tree.root().children(0) , "-" , 2 , 2 , 1 );
@@ -108,9 +108,9 @@ TEST( TESTNAME , insert_cursor )
     test_tree< basic_tree_tag >::tree_type::cursor c2 = trees.data2.root();
     trees.data.insert_below( c , c2 );
     
-    EXPECT_EQ( trees.data.size() , 9 );
+    EXPECT_EQ( trees.data.size() , size_t( 9 ) );
     EXPECT_FALSE( trees.data.empty() );
-    EXPECT_EQ( trees.data2.size() , 4 );
+    EXPECT_EQ( trees.data2.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data2.empty() );
     
     TEST_NODE( trees.data.root() , "plus" , 2 , 5 , 0 );
@@ -130,7 +130,7 @@ TEST( TESTNAME , assign )
     test_tree< basic_tree_tag > trees;
     trees.data.assign( trees.data2.root() );
     
-    EXPECT_EQ( trees.data.size() , 4 );
+    EXPECT_EQ( trees.data.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data.empty() );
     TEST_NODE( trees.data.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( trees.data.root().children(0) , "cos" , 1 , 2 , 1 );
@@ -143,7 +143,7 @@ TEST( TESTNAME , clear )
     test_tree< basic_tree_tag > trees;
     trees.data.clear();
     
-    EXPECT_EQ( trees.data.size() , 0 );
+    EXPECT_EQ( trees.data.size() , size_t( 0 ) );
     EXPECT_TRUE( trees.data.empty() );
 }
 
@@ -152,7 +152,7 @@ TEST( TESTNAME , iterator_construct )
     test_tree< basic_tree_tag > trees;
     test_tree< basic_tree_tag >::tree_type t( trees.data2.root() );
     
-    EXPECT_EQ( t.size() , 4 );
+    EXPECT_EQ( t.size() , size_t( 4 ) );
     EXPECT_FALSE( t.empty() );
     TEST_NODE( t.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( t.root().children(0) , "cos" , 1 , 2 , 1 );
@@ -165,9 +165,9 @@ TEST( TESTNAME , copy_construct )
     test_tree< basic_tree_tag > trees;
     test_tree< basic_tree_tag >::tree_type t( trees.data2 );
     
-    EXPECT_EQ( trees.data2.size() , 4 );
+    EXPECT_EQ( trees.data2.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data2.empty() );
-    EXPECT_EQ( t.size() , 4 );
+    EXPECT_EQ( t.size() , size_t( 4 ) );
     EXPECT_FALSE( t.empty() );
     TEST_NODE( t.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( t.root().children(0) , "cos" , 1 , 2 , 1 );
@@ -180,10 +180,10 @@ TEST( TESTNAME , move_construct )
     test_tree< basic_tree_tag > trees;
     test_tree< basic_tree_tag >::tree_type t( std::move( trees.data2 ) );
     
-    EXPECT_EQ( trees.data2.size() , 0 );
+    EXPECT_EQ( trees.data2.size() , size_t( 0 ) );
     EXPECT_TRUE( trees.data2.empty() );
     
-    EXPECT_EQ( t.size() , 4 );
+    EXPECT_EQ( t.size() , size_t( 4 ) );
     EXPECT_FALSE( t.empty() );
     TEST_NODE( t.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( t.root().children(0) , "cos" , 1 , 2 , 1 );
@@ -196,10 +196,10 @@ TEST( TESTNAME , copy_assign )
     test_tree< basic_tree_tag > trees;
     trees.data = trees.data2;
     
-    EXPECT_EQ( trees.data2.size() , 4 );
+    EXPECT_EQ( trees.data2.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data2.empty() );
     
-    EXPECT_EQ( trees.data.size() , 4 );
+    EXPECT_EQ( trees.data.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data.empty() );
     TEST_NODE( trees.data.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( trees.data.root().children(0) , "cos" , 1 , 2 , 1 );
@@ -212,10 +212,10 @@ TEST( TESTNAME , move_assign )
     test_tree< basic_tree_tag > trees;
     trees.data = std::move( trees.data2 );
     
-    EXPECT_EQ( trees.data2.size() , 0 );
+    EXPECT_EQ( trees.data2.size() , size_t( 0 ) );
     EXPECT_TRUE( trees.data2.empty() );
     
-    EXPECT_EQ( trees.data.size() , 4 );
+    EXPECT_EQ( trees.data.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data.empty() );
     TEST_NODE( trees.data.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( trees.data.root().children(0) , "cos" , 1 , 2 , 1 );
@@ -229,14 +229,14 @@ TEST( TESTNAME , swap_method )
     test_tree< basic_tree_tag > trees;
     trees.data.swap( trees.data2 );
     
-    EXPECT_EQ( trees.data.size() , 4 );
+    EXPECT_EQ( trees.data.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data.empty() );
     TEST_NODE( trees.data.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( trees.data.root().children(0) , "cos" , 1 , 2 , 1 );
     TEST_NODE( trees.data.root().children(0).children(0) , "y" , 0 , 1 , 2 );
     TEST_NODE( trees.data.root().children(1) , "x" , 0 , 1 , 1 );
 
-    EXPECT_EQ( trees.data2.size() , 6 );
+    EXPECT_EQ( trees.data2.size() , size_t( 6 ) );
     EXPECT_FALSE( trees.data2.empty() );
     TEST_NODE( trees.data2.root() , "plus" , 2 , 3 , 0 );
     TEST_NODE( trees.data2.root().children(0) , "sin" , 1 , 2 , 1 );
@@ -251,14 +251,14 @@ TEST( TESTNAME , swap_function )
     test_tree< basic_tree_tag > trees;
     swap( trees.data , trees.data2 );
     
-    EXPECT_EQ( trees.data.size() , 4 );
+    EXPECT_EQ( trees.data.size() , size_t( 4 ) );
     EXPECT_FALSE( trees.data.empty() );
     TEST_NODE( trees.data.root() , "minus" , 2 , 3 , 0 );
     TEST_NODE( trees.data.root().children(0) , "cos" , 1 , 2 , 1 );
     TEST_NODE( trees.data.root().children(0).children(0) , "y" , 0 , 1 , 2 );
     TEST_NODE( trees.data.root().children(1) , "x" , 0 , 1 , 1 );
 
-    EXPECT_EQ( trees.data2.size() , 6 );
+    EXPECT_EQ( trees.data2.size() , size_t( 6 ) );
     EXPECT_FALSE( trees.data2.empty() );
     TEST_NODE( trees.data2.root() , "plus" , 2 , 3 , 0 );
     TEST_NODE( trees.data2.root().children(0) , "sin" , 1 , 2 , 1 );
@@ -301,8 +301,8 @@ TEST( TESTNAME , swap_subtrees1 )
     swap_subtrees( t1 , t1.root() , t2 , t2.root() );
     EXPECT_TRUE( t1.empty() );
     EXPECT_TRUE( t2.empty() );
-    EXPECT_EQ( t1.size() , 0 );
-    EXPECT_EQ( t2.size() , 0 );
+    EXPECT_EQ( t1.size() , size_t( 0 ) );
+    EXPECT_EQ( t2.size() , size_t( 0 ) );
 }
 
 TEST( TESTNAME , swap_subtrees2 )
@@ -312,8 +312,8 @@ TEST( TESTNAME , swap_subtrees2 )
     swap_subtrees( t1 , t1.root() , t2 , t2.root() );
     EXPECT_TRUE( t1.empty() );
     EXPECT_FALSE( t2.empty() );
-    EXPECT_EQ( t1.size() , 0 );
-    EXPECT_EQ( t2.size() , 1 );
+    EXPECT_EQ( t1.size() , size_t( 0 ) );
+    EXPECT_EQ( t2.size() , size_t( 1 ) );
 }
 
 TEST( TESTNAME , swap_subtrees3 )
@@ -324,8 +324,8 @@ TEST( TESTNAME , swap_subtrees3 )
     swap_subtrees( t1 , t1.root() , t2 , t2.root() );
     EXPECT_TRUE( t1.empty() );
     EXPECT_FALSE( t2.empty() );
-    EXPECT_EQ( t1.size() , 0 );
-    EXPECT_EQ( t2.size() , 2 );
+    EXPECT_EQ( t1.size() , size_t( 0 ) );
+    EXPECT_EQ( t2.size() , size_t( 2 ) );
     TEST_NODE( t2.root() , "+" , 1 , 2 , 0 );
     TEST_NODE( t2.root().children(0) , "-" , 0 , 1 , 1 );
 }
@@ -338,8 +338,8 @@ TEST( TESTNAME , swap_subtrees4 )
     swap_subtrees( t1 , t1.root().children(0) , t2 , t2.root() );
     EXPECT_FALSE( t1.empty() );
     EXPECT_FALSE( t2.empty() );
-    EXPECT_EQ( t1.size() , 1 );
-    EXPECT_EQ( t2.size() , 1 );
+    EXPECT_EQ( t1.size() , size_t( 1 ) );
+    EXPECT_EQ( t2.size() , size_t( 1 ) );
     TEST_NODE( t1.root() , "+" , 0 , 1 , 0 );
     TEST_NODE( t2.root() , "-" , 0 , 1 , 0 );
 }
@@ -351,8 +351,8 @@ TEST( TESTNAME , swap_subtrees5 )
     
     EXPECT_FALSE( trees.data.empty() );
     EXPECT_FALSE( trees.data2.empty() );
-    EXPECT_EQ( trees.data.size() , 5 );
-    EXPECT_EQ( trees.data2.size() , 5 );
+    EXPECT_EQ( trees.data.size() , size_t( 5 ) );
+    EXPECT_EQ( trees.data2.size() , size_t( 5 ) );
     TEST_NODE( trees.data.root() , "plus" , 2 , 3 , 0 );
     TEST_NODE( trees.data.root().children(0) , "sin" , 1 , 2 , 1 );
     TEST_NODE( trees.data.root().children(0).children(0) , "x" , 0 , 1 , 2 );
